requirements.txt
quickfs==0.1.1
google-cloud-bigquery
google-cloud-storage
pandas